moved new api funcs back into api.go
- msg_actions has less code now
- some minor deletions in commandfuncs.go for rpc, should be ignored.
